+++ lib/src/codecs/base16.dart
@@ -5,24 +5,20 @@ import 'codec.dart';
 import 'converter.dart';
 
 const int _zero = 48;
-const int _smallA = 97;
 const int _bigA = 65;
+const int _smallA = 97;
 
-const _base16Decoding = <int>[
-  -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, //
-  -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2,
-  -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, -2, -2,
-  -2, -2, -2, -2, -2, 10, 11, 12, 13, 14, 15, -2, -2, -2, -2, -2, -2, -2, -2,
-  -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, 10,
-  11, 12, 13, 14, 15,
-];
+class _B16Encoder extends Uint8Encoder {
+  final int startCode;
 
-class _B16UpperEncoder extends Uint8Encoder {
-  const _B16UpperEncoder()
+  const _B16Encoder._(this.startCode)
       : super(
           bits: 4,
           alphabet: const <int>[],
         );
+
+  static const upper = _B16Encoder._(_bigA - 10);
+  static const lower = _B16Encoder._(_smallA - 10);
 
   @override
   Iterable<int> convert(Iterable<int> input) sync* {
@@ -30,29 +26,8 @@ class _B16UpperEncoder extends Uint8Encoder {
     for (int x in input) {
       a = (x >>> 4) & 0xF;
       b = x & 0xF;
-      a += a < 10 ? _zero : (_bigA - 10);
-      b += b < 10 ? _zero : (_bigA - 10);
-      yield a;
-      yield b;
-    }
-  }
-}
-
-class _B16LowerEncoder extends Uint8Encoder {
-  const _B16LowerEncoder()
-      : super(
-          bits: 4,
-          alphabet: const <int>[],
-        );
-
-  @override
-  Iterable<int> convert(Iterable<int> input) sync* {
-    int a, b, x;
-    for (x in input) {
-      a = (x >>> 4) & 0xF;
-      b = x & 0xF;
-      a += a < 10 ? _zero : (_smallA - 10);
-      b += b < 10 ? _zero : (_smallA - 10);
+      a += a < 10 ? _zero : startCode;
+      b += b < 10 ? _zero : startCode;
       yield a;
       yield b;
     }
@@ -63,7 +38,7 @@ class _B16Decoder extends Uint8Decoder {
   const _B16Decoder()
       : super(
           bits: 4,
-          alphabet: _base16Decoding,
+          alphabet: const <int>[],
         );
 
   @override
@@ -74,7 +49,16 @@ class _B16Decoder extends Uint8Decoder {
     n = input.length;
     f = (n & 1 != 0);
     for (y in input) {
-      if (y >= alphabet.length || (x = alphabet[y]) < 0) {
+      if (y >= _smallA) {
+        x = y - _smallA + 10;
+      } else if (y >= _bigA) {
+        x = y - _bigA + 10;
+      } else if (y >= _zero) {
+        x = y - _zero;
+      } else {
+        x = -1;
+      }
+      if (x < 0 || x > 15) {
         throw FormatException('Invalid character $y');
       }
       if (f) {
@@ -95,10 +79,10 @@ class B16Codec extends Uint8Codec {
   final decoder = const _B16Decoder();
 
   /// Base16 codec with lowercase letters
-  const B16Codec() : encoder = const _B16UpperEncoder();
+  const B16Codec() : encoder = _B16Encoder.upper;
 
   /// Base16 codec with uppercase letters
-  const B16Codec.lower() : encoder = const _B16LowerEncoder();
+  const B16Codec.lower() : encoder = _B16Encoder.lower;
 }
 
 /// Codec to encode and decode an iterable of 8-bit integers to 4-bit Base16
@@ -120,12 +104,25 @@ const base16 = B16Codec();
 const base16lower = B16Codec.lower();
 
 /// Encode an array of 8-bit integers to Base16 (hexadecimal) string
-String toHex(Iterable<int> input, {bool upper = false}) {
+///
+/// Parameters:
+/// - If [upper] is true, the string will be in uppercase alphabets.
+/// - If [padding] is true, the string will be padded with 0 at the start.
+String toHex(
+  Iterable<int> input, {
+  bool upper = false,
+  bool padding = true,
+}) {
+  Iterable<int> out;
   if (upper) {
-    return String.fromCharCodes(base16.encoder.convert(input));
+    out = base16.encoder.convert(input);
   } else {
-    return String.fromCharCodes(base16lower.encoder.convert(input));
+    out = base16lower.encoder.convert(input);
   }
+  if (!padding) {
+    out = out.skipWhile((value) => value == _zero);
+  }
+  return String.fromCharCodes(out);
 }
 
 /// Decode an array of 8-bit integers from Base16 (hexadecimal) string